+++ packages/moneystream-extension/src/content/util/getAdaptedSite.ts
@@ -1,6 +1,7 @@
 export const ADAPTED_REGEX: Record<string, RegExp> = {
   twitch: /https?:\/\/(www\.)?twitch\.tv(\/.*)?/i,
-  youtube: /https?:\/\/(www\.)?youtube\.com\/(watch|channel)(.*)/i
+  youtube: /https?:\/\/(www\.)?youtube\.com\/(watch|channel)(.*)/i,
+  powping: /https?:\/\/(www\.)?powping\.com\/posts\/[0-9a-fA-F]{64}?/i,
 }
 
 const ADAPTED_SITES = Object.keys(ADAPTED_REGEX)